pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.27.0",
    "numpy>=2.3.2",
    "sift-stack-py>=0.7.0",
]
